requirements.txt
# Core dependencies
fire==0.7.0
PyYAML==6.0.2
transformers==4.57.3
accelerate

# traitlens dependencies (now bundled in traitlens/)
scipy>=1.7.0
scikit-learn>=1.0.0

# LLM-as-judge classification (for deconfounded extraction)
google-genai
python-dotenv
openai
